+++ arrus/core/devices/us4r/Us4RImpl.cpp
@@ -277,7 +277,7 @@ void Us4RImpl::setActiveTermination(std::optional<uint16> value) {
 }
 
 uint8_t Us4RImpl::getNumberOfUs4OEMs() {
-    return us4oems.size();
+    return static_cast<uint8_t>(us4oems.size());
 }
 
 float Us4RImpl::getSamplingFrequency() const {
