+++ arrus/core/devices/us4r/probeadapter/ProbeAdapterImplTest.cpp
@@ -93,7 +93,7 @@ public:
          bool triggerSync),
         (override));
     MOCK_METHOD(Interval<Voltage>, getAcceptedVoltageRange, (), (override));
-    MOCK_METHOD(double, getSamplingFrequency, (), (override));
+    MOCK_METHOD(float, getSamplingFrequency, (), (override));
     MOCK_METHOD(float, getFPGATemperature, (), (override));
     MOCK_METHOD(void, startTrigger, (), (override));
     MOCK_METHOD(void, stopTrigger, (), (override));
